requirements.txt
pyrogram
tgcrypto
uvloop

python-dotenv
python-decouple

tortoise-orm
asyncpg
aerich
asgiref
flask